Avoid possible overflow in socket_putmessage_nonblock
On systems where int consists of fewer bits than size_t there was a
possibility for overflow in socket_putmessage_nonblock, because of an
intermediate cast from size_t to int.

+++ src/backend/libpq/pqcomm.c
@@ -1487,7 +1487,7 @@ static void
 socket_putmessage_noblock(char msgtype, const char *s, size_t len)
 {
 	int			res PG_USED_FOR_ASSERTS_ONLY;
-	int			required;
+	size_t		required;
 
 	/*
 	 * Ensure we have enough space in the output buffer for the message header
